[DEV-3014] test adaption and minor refactoring
