+++ src/subdomains/core/payment-link/entities/standard.config.ts
@@ -1,11 +1,5 @@
 import { Blockchain } from 'src/integration/blockchain/shared/enums/blockchain.enum';
-
-export enum PaymentStandard {
-  OPEN_CRYPTO_PAY = 'OpenCryptoPay',
-  FRANKENCOIN_PAY = 'FrankencoinPay',
-  LIGHTNING_BOLT11 = 'LightningBolt11',
-  PAY_TO_ADDRESS = 'PayToAddress',
-}
+import { PaymentStandard } from '../enums';
 
 export interface PaymentLinkConfig {
   standards: PaymentStandard[];
