+++ src/subdomains/generic/forwarding/controllers/__tests__/lnurl-forward.spec.ts
@@ -65,7 +65,7 @@ describe('LnurlForward', () => {
         }),
       );
 
-      const result = await lnurlpForward.lnUrlPForward('ABC123');
+      const result = await lnurlpForward.lnUrlPForward('ABC123', undefined);
 
       expect(result).toEqual(
         createCustomLnurlpLRequest({
